Expanded Acquia API endpoints
