+++ src/Blt/Plugin/Commands/AliasesCommand.php
@@ -16,34 +16,6 @@ class AliasesCommand extends BltTasks {
   use SwsCommandTrait;
 
   /**
-   * App id.
-   *
-   * @var string
-   */
-  protected $appId;
-
-  /**
-   * Cloud config dir.
-   *
-   * @var string
-   */
-  protected $cloudConfDir;
-
-  /**
-   * Cloud config filename.
-   *
-   * @var string
-   */
-  protected $cloudConfFileName;
-
-  /**
-   * Cloud config file path.
-   *
-   * @var string
-   */
-  protected $cloudConfFilePath;
-
-  /**
    * Site alias dir.
    *
    * @var string
@@ -58,14 +30,8 @@ class AliasesCommand extends BltTasks {
    * @aliases raia aliases
    */
   public function generateAliasesAcquia() {
-    $this->cloudConfDir = $_SERVER['HOME'] . '/.acquia';
-    $this->setAppId();
-    $this->cloudConfFileName = 'cloud_api.conf';
-    $this->cloudConfFilePath = $this->cloudConfDir . '/' . $this->cloudConfFileName;
+    $this->connectAcquiaApi();
     $this->siteAliasDir = $this->getConfigValue('drush.alias-dir');
-
-    $cloudApiConfig = $this->loadCloudApiConfig();
-    $this->setCloudApiClient($cloudApiConfig['key'], $cloudApiConfig['secret']);
 
     $this->say("<info>Gathering site info from Acquia Cloud.</info>");
     $site = $this->acquiaApplications->get($this->appId);
@@ -83,108 +49,6 @@ class AliasesCommand extends BltTasks {
     }
   }
 
-  /**
-   * Sets the Acquia application ID from config and prompt.
-   */
-  protected function setAppId() {
-    if ($app_id = $this->getConfigValue('cloud.appId')) {
-      $this->appId = $app_id;
-    }
-    else {
-      $this->say("<info>To generate an alias for the Acquia Cloud, BLT requires your Acquia Cloud application ID.</info>");
-      $this->say("<info>See https://docs.acquia.com/acquia-cloud/manage/applications.</info>");
-      $this->appId = $this->askRequired('Please enter your Acquia Cloud application ID');
-      $this->writeAppConfig($this->appId);
-    }
-  }
-
-  /**
-   * Sets appId value in blt.yml to disable interative prompt.
-   *
-   * @param string $app_id
-   *   The Acquia Cloud application UUID.
-   *
-   * @throws \Acquia\Blt\Robo\Exceptions\BltException
-   */
-  protected function writeAppConfig($app_id) {
-
-    $project_yml = $this->getConfigValue('blt.config-files.project');
-    $this->say("Updating ${project_yml}...");
-    $project_config = YamlMunge::parseFile($project_yml);
-    $project_config['cloud']['appId'] = $app_id;
-    try {
-      YamlMunge::writeFile($project_yml, $project_config);
-    }
-    catch (\Exception $e) {
-      throw new BltException("Unable to update $project_yml.");
-    }
-  }
-
-  /**
-   * Loads CloudAPI token from an user input if it doesn't exist on disk.
-   *
-   * @return array
-   *   An array of CloudAPI token configuration.
-   */
-  protected function loadCloudApiConfig() {
-    if (!$config = $this->loadCloudApiConfigFile()) {
-      $config = $this->askForCloudApiCredentials();
-    }
-    return $config;
-  }
-
-  /**
-   * Load existing credentials from disk.
-   *
-   * @return bool|array
-   *   Returns credentials as array on success, or FALSE on failure.
-   */
-  protected function loadCloudApiConfigFile() {
-    if (file_exists($this->cloudConfFilePath)) {
-      return (array) json_decode(file_get_contents($this->cloudConfFilePath));
-    }
-    else {
-      return FALSE;
-    }
-  }
-
-  /**
-   * Interactive prompt to get Cloud API credentials.
-   *
-   * @return array
-   *   Returns credentials as array on success.
-   *
-   * @throws \Acquia\Blt\Robo\Exceptions\BltException
-   */
-  protected function askForCloudApiCredentials() {
-    $this->say("You may generate new API tokens at <comment>https://cloud.acquia.com/app/profile/tokens</comment>");
-    $key = $this->askRequired('Please enter your Acquia cloud API key:');
-    $secret = $this->askRequired('Please enter your Acquia cloud API secret:');
-
-    // Attempt to set client to check credentials (throws exception on failure).
-    $this->setCloudApiClient($key, $secret);
-
-    $config = [
-      'key' => $key,
-      'secret' => $secret,
-    ];
-    $this->writeCloudApiConfig($config);
-    return $config;
-  }
-
-  /**
-   * Writes configuration to local file.
-   *
-   * @param array $config
-   *   An array of CloudAPI configuraton.
-   */
-  protected function writeCloudApiConfig(array $config) {
-    if (!is_dir($this->cloudConfDir)) {
-      mkdir($this->cloudConfDir);
-    }
-    file_put_contents($this->cloudConfFilePath, json_encode($config));
-    $this->say("Credentials were written to {$this->cloudConfFilePath}.");
-  }
 
   /**
    * Gets generated drush site aliases.
@@ -255,7 +119,8 @@ class AliasesCommand extends BltTasks {
                 $sites[$siteID][$envName] = ['uri' => $uri];
                 $siteAlias = $this->getAliases($uri, $envName, $remoteHost, $remoteUser, $siteID);
                 $sites[$siteID][$envName] = $siteAlias[$envName];
-              } continue;
+              }
+              continue;
             }
           }
 
@@ -324,7 +189,6 @@ class AliasesCommand extends BltTasks {
 
     $this->say('Getting ACSF sites.json information...');
     $result = $this->taskRsync()
-
       ->fromPath('/mnt/files/' . $remoteUser . '/files-private/sites.json')
       ->fromHost($sshFull)
       ->toPath($this->cloudConfDir)
